+++ core/src/main/scala/me/maciejb/snappyflows/impl/Int24.scala
@@ -1,4 +1,4 @@
-package me.maciejb.snappyflows.util
+package me.maciejb.snappyflows.impl
 
 import akka.stream.io.ByteStringParser.ByteReader
 import akka.util.ByteString
